requirements.txt
kafka-python==2.0.2
requests==2.25.0
psycopg2-binary==2.8.6
